# server/docs/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "PageMD EMR API",
        "version": "1.0.0",
        "description": "Commercial-grade REST API for PageMD Electronic Medical Records",
        "contact": {
            "name": "PageMD Support",
            "url": "https://pagemdemr.com"
        },
        "license": {
            "name": "Proprietary",
            "url": "https://pagemdemr.com/terms"
        }
    },
    "servers": [
        {
            "url": "https://api.pagemdemr.com",
            "description": "Production"
        },
        {
            "url": "https://sandbox.pagemdemr.com",
            "description": "Sandbox"
        }
    ],
    "security": [
        {
            "oauth2": []
        }
    ],
...
    "paths": {
...
        "/api/v1/webhooks/subscriptions": {
            "get": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "List webhook subscriptions",
                "operationId": "listWebhookSubscriptions",
                "responses": {
                    "200": {
                        "description": "List of subscriptions"
                    }
                },
                "security": [
                    {
                        "oauth2": [
                            "webhook.manage"
                        ]
                    }
                ]
            },
            "post": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Create webhook subscription",
                "operationId": "createWebhookSubscription",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/WebhookSubscriptionCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Subscription created with secret"
                    }
                },
                "security": [
                    {
                        "oauth2": [
                            "webhook.manage"
                        ]
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "oauth2": {
                "type": "oauth2",
                "flows": {
                    "authorizationCode": {
                        "authorizationUrl": "/oauth/authorize",
                        "tokenUrl": "/oauth/token",
                        "scopes": {
                            "patient.read": "Read patient records",
                            "patient.write": "Create/update patients",
                            "appointment.read": "Read appointments",
                            "appointment.write": "Create/update appointments",
                            "encounter.read": "Read encounters",
                            "encounter.write": "Create/update encounters",
                            "document.read": "Read documents",
                            "document.write": "Upload documents",
                            "webhook.manage": "Manage webhooks",
                            "ai.use": "Use AI features"
                        }
                    },
                    "clientCredentials": {
                        "tokenUrl": "/oauth/token",
                        "scopes": {}
                    }
                }
            },
            "basicAuth": {
                "type": "http",
                "scheme": "basic"
            }
        },
        "schemas": {
...
            "WebhookSubscriptionCreate": {
                "type": "object",
                "properties": {
                    "name": {
                        "type": "string"
                    },
                    "url": {
                        "type": "string",
                        "format": "uri"
                    },
                    "events": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                },
                "required": [
                    "url",
                    "events"
                ]
            },
...
        },
...
    }
}